refactor: sequenceRevealPanel into 4 files
